+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Version 0.3
 
+- [ ] Add method to toggle list ordered types:
+  - [x] For normal mode
+  - [ ] For visual mode
 - [ ] Handle multi-line bullets
 - [ ] Handle checking task markers
 - [ ] Handle insertion with "O"